requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.2.0
cycler==0.12.1
dgl==1.1.2+cu118
filelock==3.12.4
fonttools==4.45.1
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
imbalanced-learn==0.11.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.2.2
Pillow==10.0.1
psutil==5.9.6
pygraphviz==1.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-swats==0.1.0
pytorch_optimizer==2.12.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
triton==2.1.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.0.6